refactor: extract shared logic from game data hooks into utility modules
Move duplicated buildPlayerIndex, buildGameScoresMap, and
buildSeasonWhereClause into lib/gameDataUtils.ts and lib/queryUtils.ts.
Both NBA and NHL hooks now use the shared implementations, with
playoffTypes parameter handling the NBA vs NHL difference.
